feat(colorschemes): Persist color picker choices (#33)
Previously, color scheme picker choices persisted only per-session. This
meant I never bothered using the picker and even manually forked these
dotfiles to my `cosgroveb` EMU account.

Now your choice persists for the life of your cpair.

A future enhancement could be to persist the `telescope-colorscheme-persist`
`file_path`[1] to `~/bt/neovim-dotfiles-personal` where it can be
committed so new cpairs get it. It could be persisted to atuin or s3?
Probably git, though.

Huuge H/T to @diecarrillo for finding this.

With this, I think we can reasonably make the legacy vim-dotfiles scheme
the default (#16) in order to increase nvim adoption.

This closes #25.

[1]: https://github.com/tingey21/telescope-colorscheme-persist.nvim/?tab=readme-ov-file#default-configuration

* Prefer opts instead of config for lazy.nvim

Co-authored-by: Hector Rios <that.hector@gmail.com>

* Re-add prematurely removed preview-enabled color scheme mapping

---------

Co-authored-by: Hector Rios <that.hector@gmail.com>