feat(sessions): minimal, non-intrusive sess. mgmt (#32)
After having some frustration with the auto-session API in some personal
work, I took a look around and found that there seems to be more
community momentum behind `resession`. It's the default in `AstroNVim`
and in addition to solving some personal frustrations it seems to have a
philosophy of doing nothing magically out of the box (auto-session
emphasizes the "auto").

In agreement with feedback from @kboschert and @terin this
minimal setup:

1. Does not auto save/auto restore (though personal dotfiles can tweak
   that once #1 is solved)
2. Adds three keymaps which manipulate sessions named with default logic
   (pwd:git branch):
  1. `\ss` - saves session
  2. `\sl` - loads session
  3. `\sd` - deletes session

In the future we can add keymaps that manipulate manually named sessions, perhaps.