fix(ui): telescope hidden files (#105)
* fix(ui): Make ignored directories explicitly directories

Prior to this, we were hiding anything that started with `node_modules`,
or `.git` (which includes `.github`).

By explicitly making these into directories, we're now able to achieve
what we intend (and show .github files).

Additionally, we're pinning the version of Telescope to the last release
(which was a few months ago) in order to make sure that we're not bitten
by a breaking change.

* refactor(ui): Run `\cf` to format file

* fix(ui): Show part of `.git`, and new files that aren't in git