feat(legacy): default to autoread=false (#116)
Yes, this defaults to true in neovim and false in vim.

Typically, I would not want to change a neovim default however for folks
that are used to vim's behavior this will result in unexpected loss of
changes during e.g. git ops.

I say that because when this is false in either editor, the user DOES
get a prompt asking if they want to reload the file because it has
changed.

When this is set to true it silently discards your changes. Now, if one
is used to that behavior fine. If not, you will lose changes and quit in
frustration back to vim. At least you are notified and consent to your
dirty buffer being disregarded if this is set to false.

This closes #93.