Tresitter upgrade (#241)
* update plugins

* update cheatsheet

* move to new treesitter setup

mostly stolen from lazyvim

* lockfile update

* cleanup

* chore(telescope): Upgrade to latest

The telescope project appears to have abandoned using tags for the last
two years. This updates our configuration to use the latest version
that's available, which at the time of commit is a few months ago.

Since we are using a lockfile, I believe that this should be a
relatively safe upgrade.

Long term, we should evaluate other pickers that are more maintained
such as `fff.nvim`, `snacks.picker` or `fzf-lua.nvim`.

---------

Co-authored-by: Hector Rios <that.hector@gmail.com>